pyproject.toml
[tool.poetry]
name = "daily-metal-prices"
version = "0.1.0"
description = ""
authors = ["Vktor_Booker <victoraguirre.f@gmail.com>"]
readme = "README.md"
packages = [{include = "daily_metal_prices"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
sqlmodel = "^0.0.8"
python-dotenv = "^0.21.0"
pre-commit = "^2.21.0"
pydantic = "^1.10.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
